feature: divided logic in ParkingBottomSheet file
