--- src/features/NearestParkingModal/index.ts
@@ -1,1 +1,0 @@
-export { NearestParkingModal } from "./ui/NearestParkingModal";
