+++ src/features/ParkingBottomSheet/components/ui/CustomHandle/CustomHandle.tsx
@@ -1,0 +1,8 @@
+import { View } from "react-native";
+import { styles } from "./CustomHandle.styles";
+
+export const CustomHandle = () => (
+    <View style={styles.customHandle}>
+        <View style={styles.customHandleIndicator} />
+    </View>
+);
